rename analyze.py to exchange_helpers.py
