+++ freqtrade/tests/test_misc.py
@@ -7,7 +7,7 @@ Unit test file for misc.py
 import datetime
 from unittest.mock import MagicMock
 
-from freqtrade.analyze import parse_ticker_dataframe
+from freqtrade.exchange.exchange_helpers import parse_ticker_dataframe
 from freqtrade.misc import (common_datearray, datesarray_to_datetimearray,
                             file_dump_json, format_ms_time, shorten_date)
 from freqtrade.optimize.__init__ import load_tickerdata_file
